UX: minor tweak to support tab
- smaller delete category button
- do not disable delete button while saving, only do nothing if you click on it, this is to avoid a flash on the button
- makes the border grey instead of blue for the sites settings emphasis

+++ frontend/discourse/admin/controllers/edit-category/tabs.js
@@ -338,6 +338,10 @@ export default class EditCategoryTabsController extends Controller {
 
   @action
   deleteCategory() {
+    if (this.deleteDisabled) {
+      return;
+    }
+
     this.set("deleting", true);
     this.dialog.deleteConfirm({
       title: i18n("category.delete_confirm"),
@@ -360,6 +364,10 @@ export default class EditCategoryTabsController extends Controller {
 
   @action
   toggleDeleteTooltip() {
+    if (this.deleteDisabled) {
+      return;
+    }
+
     this.toggleProperty("showTooltip");
   }
